disable_trackpad_keyboard.sh
Are you single and walang jowa? This can be your chance on getting your match, by cloning this repo, you can have the ability to make a cheezy moment or kilig vibes just in case the laptop of your crush (possible forever naks!) has a broken mouse-trackpad or keyboard. This repo is better than your dating app, try this one as well and who knows? Both of you will be come forever! Ayeee!